% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-chains.R
\name{bind_chains}
\alias{bind_chains}
\title{Combine MCMC objects by chains.}
\usage{
bind_chains(x, x2, ...)
}
\arguments{
\item{x}{An object.}

\item{x2}{A second object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The combined objects.
}
\description{
Combines two MCMC objects (with the same parameters and iterations) by chains.
}
\seealso{
Other {MCMC manipulations}: 
\code{\link{collapse_chains}()},
\code{\link{complete_terms}()},
\code{\link{fill_na}()},
\code{\link{split_chains}()}
}
\concept{{MCMC manipulations}}
