% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-iterations.R
\name{bind_iterations}
\alias{bind_iterations}
\title{Bind Iterations}
\usage{
bind_iterations(x, x2, ...)
}
\arguments{
\item{x}{An object.}

\item{x2}{A second object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The combined object.
}
\description{
Combines two MCMC objects (with the same parameters and chains) by iterations.
}
\seealso{
Other MCMC manipulations: 
\code{\link{bind_chains}()},
\code{\link{collapse_chains}()},
\code{\link{estimates}()},
\code{\link{split_chains}()}
}
\concept{MCMC manipulations}
