% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse-chains.R
\name{collapse_chains}
\alias{collapse_chains}
\title{Collapse Chains}
\usage{
collapse_chains(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object with one chain.
}
\description{
Collapses an MCMC object's chains into a single chain.
}
\seealso{
Other {MCMC dimensions}: 
\code{\link{nchains}()},
\code{\link{niters}()},
\code{\link{npars}()},
\code{\link{nsams}()},
\code{\link{nsims}()},
\code{\link{nterms}()},
\code{\link{split_chains}()}
}
\concept{{MCMC dimensions}}
