% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete-terms.R
\name{complete_terms}
\alias{complete_terms}
\title{Complete Terms}
\usage{
complete_terms(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object.
}
\description{
Completes an objects terms.
}
\seealso{
Other {terms}: 
\code{\link{nterms}()},
\code{\link{pars_terms}()}

Other {MCMC manipulations}: 
\code{\link{bind_chains}()},
\code{\link{collapse_chains}()},
\code{\link{fill_na}()},
\code{\link{split_chains}()}
}
\concept{{MCMC manipulations}}
\concept{{terms}}
