% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converged.R
\name{converged}
\alias{converged}
\title{Converged}
\usage{
converged(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A logical scalar indicating whether the object has converged.
}
\description{
Tests whether an object has converged.
}
\seealso{
Other convergence: 
\code{\link{converged_pars}()},
\code{\link{converged_terms}()},
\code{\link{esr}()},
\code{\link{esr_pars}()},
\code{\link{esr_terms}()},
\code{\link{rhat}()},
\code{\link{rhat_pars}()},
\code{\link{rhat_terms}()}
}
\concept{convergence}
