% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converged-pars.R
\name{converged_pars}
\alias{converged_pars}
\title{Converged Parameters}
\usage{
converged_pars(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A uniquely named logical atomic vector indicating whether
each parameter has converged.
}
\description{
Tests whether each parameter of an object has converged.
}
\examples{
converged_pars.foobar <- function(x, ...) {
  NotYetImplemented()
  # replace with code to get converged for a each term of an object of
  # class 'foobar'
}
}
\seealso{
Other convergence: 
\code{\link{converged_terms}()},
\code{\link{converged}()},
\code{\link{esr_pars}()},
\code{\link{esr_terms}()},
\code{\link{esr}()},
\code{\link{rhat_pars}()},
\code{\link{rhat_terms}()},
\code{\link{rhat}()}
}
\concept{convergence}
