% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converged-terms.R
\name{converged_terms}
\alias{converged_terms}
\title{Converged Terms}
\usage{
converged_terms(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A list of uniquely named logical objects with whether
each term has converged.
}
\description{
Tests whether each term of an object has converged.
}
\seealso{
Other convergence: 
\code{\link{converged_pars}()},
\code{\link{converged}()},
\code{\link{esr_pars}()},
\code{\link{esr_terms}()},
\code{\link{esr}()},
\code{\link{rhat_pars}()},
\code{\link{rhat_terms}()},
\code{\link{rhat}()}
}
\concept{convergence}
