% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esr-terms.R
\name{esr_terms}
\alias{esr_terms}
\title{Effective Sampling Rate for Terms}
\usage{
esr_terms(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A list of uniquely named numeric atomic objects with values between 0 and 1 indicating the effectively sampling rate for each term.
}
\description{
Calculates the effective sampling rate (\code{esr}) for each term.
}
\details{
By default
\deqn{\frac{1}{1 + 2 \sum_{k = 1}^\infty\rho_k(\theta)}}
from Brooks et al. (2011) where the infinite sum is truncated at
lag \eqn{k} when \eqn{\rho_{k+1}(\theta) < 0}.
}
\examples{
esr_terms.foobar <- function(x, ...) {
  NotYetImplemented()
  # replace with code to get esr for each term of an object of class 'foobar'
}
}
\references{
Brooks, S., Gelman, A., Jones, G.L., and Meng, X.-L. (Editors). 2011. Handbook for Markov Chain Monte Carlo. Taylor & Francis, Boca Raton.
}
\seealso{
Other convergence: 
\code{\link{converged_pars}()},
\code{\link{converged_terms}()},
\code{\link{converged}()},
\code{\link{esr_pars}()},
\code{\link{esr}()},
\code{\link{rhat_pars}()},
\code{\link{rhat_terms}()},
\code{\link{rhat}()}
}
\concept{convergence}
