% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-all.R
\name{fill_all}
\alias{fill_all}
\title{Fill All Values}
\usage{
fill_all(x, value, ...)
}
\arguments{
\item{x}{An object.}

\item{value}{A scalar of the value to replace all values with.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object.
}
\description{
Fills all of an object's (missing and non-missing) values while preserving the object's dimensionality and class.
}
\examples{
fill_all.foobar <- function(x, ...) {
  NotYetImplemented()
  # replace with code to fill_all for an object of class 'foobar'
}
}
\seealso{
Other {MCMC manipulations}: 
\code{\link{bind_chains}()},
\code{\link{collapse_chains}()},
\code{\link{estimates}()},
\code{\link{fill_na}()},
\code{\link{split_chains}()}
}
\concept{{MCMC manipulations}}
