% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-na.R
\name{fill_na}
\alias{fill_na}
\title{Fill Missing Values}
\usage{
fill_na(x, value, ...)
}
\arguments{
\item{x}{An object.}

\item{value}{A scalar of the value to replace missing values with.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object.
}
\description{
Fills an object's missing values while preserving the object's class.
}
\examples{
fill_na.foobar <- function(x, ...) {
  NotYetImplemented()
  # replace with code to fill_na for an object of class 'foobar'
}
}
\seealso{
Other MCMC manipulations: 
\code{\link{bind_chains}()},
\code{\link{collapse_chains}()},
\code{\link{estimates}()},
\code{\link{fill_all}()},
\code{\link{split_chains}()}
}
\concept{MCMC manipulations}
