% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchains.R
\name{nchains}
\alias{nchains}
\title{Number of Chains}
\usage{
nchains(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of chains.
}
\description{
Gets the number of chains
of an MCMC object.
}
\examples{
nchains.foobar <- function(x, ...) {
 NotYetImplemented()
 # replace with code to get nchains for an object of class 'foobar'
}
}
\seealso{
Other {MCMC dimensions}: 
\code{\link{niters}()},
\code{\link{npars_scalar}()},
\code{\link{npars}()},
\code{\link{nsams}()},
\code{\link{nsims}()},
\code{\link{nterms}()}
}
\concept{{MCMC dimensions}}
