% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npars.R
\name{npars}
\alias{npars}
\alias{npars.default}
\title{Number of Parameters}
\usage{
npars(x, ...)

\method{npars}{default}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of parameters.
}
\description{
Gets the number of parameters of an object.

The default methods returns the length of \code{\link[=pars]{pars()}} if none are
\code{NA}, otherwise it returns \code{NA}.
}
\examples{
npars.foobar <- function(x, ...) {
  length(pars(x, ...))
}
}
\seealso{
\code{\link[=pars]{pars()}}

Other MCMC dimensions: 
\code{\link{nchains}()},
\code{\link{niters}()},
\code{\link{nsams}()},
\code{\link{nsims}()},
\code{\link{nterms}()}

Other parameters: 
\code{\link{pars}()},
\code{\link{set_pars}()}
}
\concept{MCMC dimensions}
\concept{parameters}
