% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npars-scalar.R
\name{npars_scalar}
\alias{npars_scalar}
\title{Number of Scalar Parameters}
\usage{
npars_scalar(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of scalar parameters.
}
\description{
Gets the number of scalar parameters of an object.
}
\seealso{
\code{\link[=pars_scalar]{pars_scalar()}}

Other {MCMC dimensions}: 
\code{\link{nchains}()},
\code{\link{niters}()},
\code{\link{npars}()},
\code{\link{nsams}()},
\code{\link{nsims}()},
\code{\link{nterms}()}

Other {parameters}: 
\code{\link{npars}()},
\code{\link{pars_scalar}()},
\code{\link{pars}()},
\code{\link{set_pars}()}
}
\concept{{MCMC dimensions}}
\concept{{parameters}}
