% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npdims-terms.R
\name{npdims_terms}
\alias{npdims_terms}
\title{Term Number of Parameter Dimensions}
\usage{
npdims_terms(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A named list of integer vectors of the dimensions of each parameter for each term.
}
\description{
Gets the number of dimensions of each parameter for each term.
}
\seealso{
Other {dimensions}: 
\code{\link{dims}()},
\code{\link{ndims}()},
\code{\link{npdims}()},
\code{\link{pdims_terms}()},
\code{\link{pdims}()},
\code{\link{tindex}()}
}
\concept{{dimensions}}
