% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nterms.R
\name{nterms}
\alias{nterms}
\title{Number of Terms}
\usage{
nterms(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A integer scalar of the number of terms.
}
\description{
Gets the number of terms
of an MCMC object.
}
\examples{
nterms.foobar <- function(x, ...) {
  NotYetImplemented()
  # replace with code to get nterms for an object of class 'foobar'
}
}
\seealso{
Other MCMC dimensions: 
\code{\link{nchains}()},
\code{\link{niters}()},
\code{\link{npars_scalar}()},
\code{\link{npars}()},
\code{\link{nsams}()},
\code{\link{nsims}()}
}
\concept{MCMC dimensions}
