% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for universals Functions}
\arguments{
\item{x}{An object. See the individual methods for specifics.}

\item{...}{Other arguments passed to methods.}
}
\description{
Parameter Descriptions for universals Functions
}
