% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars-scalar.R
\name{pars_scalar}
\alias{pars_scalar}
\title{Scalar Parameter Names}
\usage{
pars_scalar(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A character vector of the names of the scalar parameters.
}
\description{
Gets the scalar parameter names.
}
\seealso{
Other {parameters}: 
\code{\link{npars_scalar}()},
\code{\link{npars}()},
\code{\link{pars}()},
\code{\link{set_pars}()}
}
\concept{{parameters}}
