% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars-terms.R
\name{pars_terms}
\alias{pars_terms}
\title{Term Parameters}
\usage{
pars_terms(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A character vector of the term parameter names.
}
\description{
Gets the name of each parameter for each term.
}
\seealso{
Other {parameters}: 
\code{\link{npars}()},
\code{\link{pars}()},
\code{\link{set_pars}()}

Other {terms}: 
\code{\link{complete_terms}()},
\code{\link{nterms}()}
}
\concept{{parameters}}
\concept{{terms}}
