% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdims-terms.R
\name{pdims_terms}
\alias{pdims_terms}
\title{Term Parameter Dimensions}
\usage{
pdims_terms(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A named integer vector of the number of dimensions of each parameter for each term.
}
\description{
Gets the dimensions of each parameter for each term.
}
\seealso{
Other {dimensions}: 
\code{\link{dims}()},
\code{\link{ndims}()},
\code{\link{npdims_terms}()},
\code{\link{npdims}()},
\code{\link{pdims}()},
\code{\link{tindex}()}
}
\concept{{dimensions}}
