% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat.R
\name{rhat}
\alias{rhat}
\title{R-hat}
\usage{
rhat(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The rhat value(s).
}
\description{
Calculates the uncorrected, unfolded, univariate,
split R-hat (potential scale reduction factor) value(s).
}
\references{
Gelman, A., and Rubin, D.B. 1992. Inference from Iterative Simulation Using Multiple Sequences. Statistical Science 7(4): 457–472.
}
