% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-chains.R
\name{split_chains}
\alias{split_chains}
\title{Split Chains}
\usage{
split_chains(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object with twice the chains and half the iterations.
}
\description{
Splits each chain in half
to double the number chains
and halve the number of iterations.
}
\seealso{
Other {MCMC dimensions}: 
\code{\link{nchains}()},
\code{\link{niters}()},
\code{\link{npars}()},
\code{\link{nsams}()},
\code{\link{nsims}()},
\code{\link{nterms}()}
}
\concept{{MCMC dimensions}}
