% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tindex.R
\name{tindex}
\alias{tindex}
\title{Term Index}
\usage{
tindex(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A named list of the index for each term.
}
\description{
Gets the index for each term of an object.
}
\details{
For example the index of \code{beta[2,1]} is \code{c(2L, 1L)}
while the index for \code{sigma} is \code{1L}.
It is useful for extracting the values of individual terms.
}
\seealso{
Other {dimensions}: 
\code{\link{dims}()},
\code{\link{ndims}()},
\code{\link{npdims}()},
\code{\link{pdims}()}
}
\concept{{dimensions}}
