% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals-package.R
\docType{package}
\name{universals-package}
\alias{universals}
\alias{universals-package}
\title{universals: S3 Generics for Bayesian Analyses}
\description{
Provides S3 generic methods and some default implementations
    for Bayesian analyses that generate Markov Chain Monte Carlo (MCMC) samples.
    The purpose of 'universals' is to reduce package dependencies and conflicts. 
    The 'nlist' package implements all the methods for its 'nlists' class.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/universals}
  \item Report bugs at \url{https://github.com/poissonconsulting/universals/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
