% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse-chains.R
\name{collapse_chains}
\alias{collapse_chains}
\title{Collapse Chains}
\usage{
collapse_chains(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object with one chain.
}
\description{
Collapses an MCMC object's chains into a single chain.
}
\examples{
collapse_chains.foobar <- function(x, ...) {
  NotYetImplemented()
  # replace with code to collapse_chains for an object of class 'foobar'
}
}
\seealso{
Other MCMC manipulations: 
\code{\link{bind_chains}()},
\code{\link{estimates}()},
\code{\link{fill_na}()},
\code{\link{split_chains}()}
}
\concept{MCMC manipulations}
