% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converged.R
\name{converged}
\alias{converged}
\title{Converged}
\usage{
converged(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A logical scalar indicating whether the object has converged.
}
\description{
Tests whether an object has converged.
}
\examples{
converged.foobar <- function(x, ...) {
 NotYetImplemented()
 # replace with code to get converged for an object of class 'foobar'
}
}
\seealso{
Other {convergence}: 
\code{\link{esr}()},
\code{\link{rhat}()}
}
\concept{{convergence}}
