% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{dims}
\alias{dims}
\title{Dimensions of an Object}
\usage{
dims(x, ...)
}
\arguments{
\item{x}{An object. See the individual methods for specifics.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer vector of the dimensions.
}
\description{
Gets the dimensions of an object.
Unlike \code{base::dim()}, dims works as intuitively expected for vectors and data.frames.
}
\section{Methods}{

To see all available methods use \code{\link[utils:methods]{utils::methods()}}.
}

\seealso{
\code{\link[base:dim]{base::dim()}}
}
