% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates}
\alias{estimates}
\title{Estimates}
\usage{
estimates(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An list of uniquely named numeric atomic objects.
}
\description{
Calculates the estimates for an MCMC object.
}
\examples{
estimates.foobar <- function(x, ...) {
  NotYetImplemented()
  # replace with code to get estimates for an object of class 'foobar'
}
}
\seealso{
Other MCMC manipulations: 
\code{\link{bind_chains}()},
\code{\link{collapse_chains}()},
\code{\link{fill_all}()},
\code{\link{fill_na}()},
\code{\link{split_chains}()}
}
\concept{MCMC manipulations}
