% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates}
\alias{estimates}
\title{Estimates}
\usage{
estimates(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A list of uniquely named numeric objects.
}
\description{
Calculates the estimates for an MCMC object.
}
\seealso{
Other MCMC manipulations: 
\code{\link{bind_chains}()},
\code{\link{bind_iterations}()},
\code{\link{collapse_chains}()},
\code{\link{split_chains}()}
}
\concept{MCMC manipulations}
