% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchains.R
\name{nchains}
\alias{nchains}
\title{Number of MCMC Chains}
\usage{
nchains(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of MCMC chains.
}
\description{
Gets the number of chains
of a Monte Carlo Markov Chain (MCMC) object.
}
\seealso{
Other {MCMC dimensions}: 
\code{\link{niters}()},
\code{\link{npars}()},
\code{\link{nsams}()},
\code{\link{nsims}()},
\code{\link{nterms}()}
}
\concept{{MCMC dimensions}}
