% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndims.R
\name{ndims}
\alias{ndims}
\title{Number of Dimensions}
\usage{
ndims(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A integer scalar of the number of dimensions.
}
\description{
Gets the number of dimensions of an object as returned by \code{dims()}.
}
\examples{
ndims.foobar <- function(x, ...) {
  length(dims(x, ...))
}
}
\seealso{
Other {dimensions}: 
\code{\link{dims}()},
\code{\link{npdims}()},
\code{\link{pdims}()}
}
\concept{{dimensions}}
