% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niters.R
\name{niters}
\alias{niters}
\title{Number of Iterations}
\usage{
niters(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of iterations.
}
\description{
Gets the number of iterations (in a chain)
of an MCMC object.
}
\seealso{
Other {MCMC dimensions}: 
\code{\link{nchains}()},
\code{\link{npars}()},
\code{\link{nsams}()},
\code{\link{nsims}()},
\code{\link{nterms}()}
}
\concept{{MCMC dimensions}}
