% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npars.R
\name{npars}
\alias{npars}
\title{Number of Parameters}
\usage{
npars(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of parameters.
}
\description{
Gets the number of parameters of an object.
}
\examples{
npars.foobar <- function(x, ...) {
  length(pars(x, ...))
}
}
\seealso{
\code{\link[=pars]{pars()}}

Other {MCMC dimensions}: 
\code{\link{nchains}()},
\code{\link{niters}()},
\code{\link{npars_scalar}()},
\code{\link{nsams}()},
\code{\link{nsims}()},
\code{\link{nterms}()}

Other {parameters}: 
\code{\link{npars_scalar}()},
\code{\link{pars_scalar}()},
\code{\link{pars}()},
\code{\link{set_pars}()}
}
\concept{{MCMC dimensions}}
\concept{{parameters}}
