% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npars.R
\name{npars}
\alias{npars}
\title{Number of Parameters}
\usage{
npars(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of parameters.
}
\description{
Gets the number of parameters of an object as returned by \code{\link[=pars]{pars()}}.
}
\seealso{
\code{\link[=pars]{pars()}}

Other {MCMC dimensions}: 
\code{\link{nchains}()},
\code{\link{nsims}()}
}
\concept{{MCMC dimensions}}
