% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npdims.R
\name{npdims}
\alias{npdims}
\title{Number of Parameter Dimensions}
\usage{
npdims(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A named integer vector of the number of dimensions of each parameter.
}
\description{
Gets the number of the dimensions of each parameter of an object.
}
\seealso{
Other {dimensions}: 
\code{\link{dims}()},
\code{\link{ndims}()},
\code{\link{npdims_terms}()},
\code{\link{pdims_terms}()},
\code{\link{pdims}()},
\code{\link{tindex}()}
}
\concept{{dimensions}}
