% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsams.R
\name{nsams}
\alias{nsams}
\title{Number of Samples}
\usage{
nsams(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of samples.
}
\description{
Gets the number of sample values (simulations * terms)
of an MCMC object.
}
\seealso{
Other {MCMC dimensions}: 
\code{\link{nchains}()},
\code{\link{niters}()},
\code{\link{npars_scalar}()},
\code{\link{npars}()},
\code{\link{nsims}()},
\code{\link{nterms}()}
}
\concept{{MCMC dimensions}}
