% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nterms.R
\name{nterms}
\alias{nterms}
\title{Number of MCMC Terms}
\usage{
nterms(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A count of the number of terms.
}
\description{
Gets the number of terms
of a Monte Carlo Markov Chain (MCMC) object.
}
\seealso{
Other {MCMC dimensions}: 
\code{\link{nchains}()},
\code{\link{niters}()},
\code{\link{npars}()},
\code{\link{nsams}()},
\code{\link{nsims}()}

Other {terms}: 
\code{\link{pars_terms}()}
}
\concept{{MCMC dimensions}}
\concept{{terms}}
