% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nterms.R
\name{nterms}
\alias{nterms}
\title{Number of Terms}
\usage{
nterms(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A integer scalar of the number of terms.
}
\description{
Gets the number of terms
of an MCMC object.
}
\seealso{
Other {MCMC dimensions}: 
\code{\link{nchains}()},
\code{\link{niters}()},
\code{\link{npars}()},
\code{\link{nsams}()},
\code{\link{nsims}()}
}
\concept{{MCMC dimensions}}
