% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{pars}
\alias{pars}
\title{Get Parameters}
\usage{
pars(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A character vector of the parameter names.
}
\description{
Gets character vector of parameter names.
}
\seealso{
Other {parameters}: 
\code{\link{npars}()},
\code{\link{set_pars}()}
}
\concept{{parameters}}
