% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{pars}
\alias{pars}
\title{Parameter Names}
\usage{
pars(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A character vector of the names of the parameters.
}
\description{
Gets the parameter names.
}
\seealso{
Other parameters: 
\code{\link{npars}()},
\code{\link{set_pars}()}
}
\concept{parameters}
