% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{pars}
\alias{pars}
\title{Get Parameters}
\usage{
pars(x, ...)
}
\arguments{
\item{x}{An object. See the individual methods for specifics.}

\item{...}{Other arguments passed to methods.}
}
\description{
Gets character vector of parameter names.
}
\section{Methods}{

To see all available methods type \verb{[methods]("pars")}.
}

