% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{pars}
\alias{pars}
\title{Parameter Names}
\usage{
pars(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A character vector of the names of the parameters.
}
\description{
Gets the parameter names.
}
\examples{
pars.foobar <- function(x, ...) {
 NotYetImplemented()
 # replace with code to get pars for an object of class 'foobar'
}
}
\seealso{
Other {parameters}: 
\code{\link{npars_scalar}()},
\code{\link{npars}()},
\code{\link{pars_scalar}()},
\code{\link{set_pars}()}
}
\concept{{parameters}}
