% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdims.R
\name{pdims}
\alias{pdims}
\title{Parameter Dimensions}
\usage{
pdims(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A named list of the dimensions of each parameter.
}
\description{
Gets the parameter dimensions of an object.
}
\seealso{
Other {dimensions}: 
\code{\link{dims}()},
\code{\link{ndims}()}
}
\concept{{dimensions}}
