% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-chains.R
\name{split_chains}
\alias{split_chains}
\title{Split Chains}
\usage{
split_chains(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object.
}
\description{
Splits each of an MCMC object's chains in half
to double the number of chains
and halve the number of iterations.
}
\seealso{
Other MCMC manipulations: 
\code{\link{bind_chains}()},
\code{\link{collapse_chains}()},
\code{\link{estimates}()}
}
\concept{MCMC manipulations}
