% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-chains.R
\name{split_chains}
\alias{split_chains}
\title{Split Chains}
\usage{
split_chains(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object.
}
\description{
Splits each of an MCMC object's chains in half
to double the number chains
and halve the number of iterations.
}
\examples{
split_chains.foobar <- function(x, ...) {
  NotYetImplemented()
  # replace with code to split_chains for an object of class 'foobar'
}
}
\seealso{
Other {MCMC manipulations}: 
\code{\link{bind_chains}()},
\code{\link{collapse_chains}()},
\code{\link{estimates}()},
\code{\link{fill_all}()},
\code{\link{fill_na}()}
}
\concept{{MCMC manipulations}}
