% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals-package.R
\docType{package}
\name{universals-package}
\alias{universals}
\alias{universals-package}
\title{universals: More Common Model-Fitting S3 Generics}
\description{
Provides common model-fitting S3 generic methods 
    missing from base R and the generics package to reduce package dependencies
    and conflicts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/universals}
  \item Report bugs at \url{https://github.com/poissonconsulting/universals/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
