% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals-package.R
\docType{package}
\name{universals-package}
\alias{universals}
\alias{universals-package}
\title{universals: Common S3 Generics for Bayesian Analyses}
\description{
Provides common S3 generic methods for Bayesian Analyses
    that are missing from base R and the generics package 
    to reduce package dependencies and conflicts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/universals}
  \item Report bugs at \url{https://github.com/poissonconsulting/universals/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
